Refactor application structure: implement FastAPI app, MongoDB connection, data access layer, and processors; add deployment and service configurations for Kubernetes.
