+++ Dockerfile
@@ -6,6 +6,8 @@ COPY requirements.txt .
 
 RUN pip install --no-cache-dir -r requirements.txt
 
-COPY hostile-tweets-ex ./hostile-tweets-ex
+COPY app ./app
+COPY data ./data
 
-CMD ["uvicorn", "hostile-tweets-ex.app.main:app", "--host", "0.0.0.0", "--port", "8000"]
+
+CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
