+++ app/processors/__init__.py
@@ -1,0 +1,1 @@
+# Init file for processors package
