+++ app/processors/weapon_processor.py
