+++ data/weapon_list.txt
