+++ infra/app-route.yaml
@@ -1,0 +1,12 @@
+apiVersion: route.openshift.io/v1
+kind: Route
+metadata:
+  name: hostile-tweets-route
+spec:
+  to:
+    kind: Service
+    name: hostile-tweets-service
+  port:
+    targetPort: 8000
+  tls:
+    termination: edge
