+++ infra/mongo-secret.yaml
@@ -1,0 +1,8 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: mongo-secret
+type: Opaque
+stringData:
+  MONGODB_CONNECTION_STRING: "mongodb+srv://IRGC:iraniraniran@iranmaldb.gurutam.mongodb.net/"
+  MONGODB_DBNAME: "IranMalDB"
