+++ scripts/commands.bat
@@ -1,0 +1,2 @@
+docker build -t chaimshvadron/malicious_text_feature:V1
+
